+++ retroarch.c
@@ -1261,7 +1261,7 @@ static size_t find_driver_nonempty(
 
 int driver_find_index(const char *label, const char *drv)
 {
-   size_t i;
+   int i;
    char str[NAME_MAX_LENGTH];
 
    str[0] = '\0';
@@ -1288,7 +1288,7 @@ int driver_find_index(const char *label, const char *drv)
  **/
 static void driver_find_last(const char *label, char *s, size_t len)
 {
-   size_t i;
+   int i;
    for (i = 0;
          find_driver_nonempty(label, i, s, len) > 0; i++) { }
    if (i)
@@ -1805,7 +1805,7 @@ void drivers_init(
 #endif
 
 #ifdef HAVE_MENU
-   srand(time(NULL));
+   srand((unsigned)time(NULL));
 #endif
 }
 
@@ -5408,7 +5408,7 @@ bool command_event(enum event_command cmd, void *data)
          break;
       case CMD_EVENT_RUMBLE_STOP:
          {
-            size_t i;
+            unsigned i;
             for (i = 0; i < MAX_USERS; i++)
             {
                unsigned joy_idx = settings->uints.input_joypad_index[i];
